sell: pan/bank flow from locker

+++ react_webviews/src/gold/components/ui_components/onload_and_timer.js
@@ -93,6 +93,10 @@ class GoldOnloadAndTimerClass extends Component {
 
     onload() {
 
+        if(this.props.parent.state.pan_bank_flow) {
+            return;
+        }
+        
         storageService().remove('forceBackState');
 
         let orderData = storageService().getObject(this.state.orderKey);
